moved home components to components directory
