+++ src/components/home/OfferApproach.tsx
